requirements.txt
boto3
python-dotenv
mangum
fastapi==0.95.1
pydantic==1.10.7
tensorflow==2.12.0
keras==2.12.0
deepface==0.0.84
Pillow>=9.0.0
faiss-cpu>=1.7.4
numpy>=1.21.0
redis>=4.5.0
h5py>=3.8.0
scikit-learn>=1.0.0
tqdm>=4.64.0